+++ msvc/mono.def
@@ -1,9 +1,9 @@
 ; file generated by create-windef.pl
 LIBRARY mono-2.0.dll
 EXPORTS
+g_mem_set_vtable
 MonoFixupCorEE
 mono_add_internal_call
-mono_set_allocator_vtable
 mono_aot_get_method
 mono_aot_register_module
 mono_array_addr_with_size
